feat(sse): SetLastEventID
