+++ pkg/protocol/sse/utils.go
@@ -17,14 +17,40 @@
 package sse
 
 import (
-	"github.com/cloudwego/hertz/pkg/app"
+	"bytes"
+
+	"github.com/cloudwego/hertz/internal/bytestr"
+	"github.com/cloudwego/hertz/pkg/protocol"
 )
 
 const LastEventIDHeader = "Last-Event-ID"
 
 // GetLastEventID returns the value of the Last-Event-ID header.
-func GetLastEventID(c *app.RequestContext) string {
-	return string(c.GetHeader(LastEventIDHeader))
+func GetLastEventID(req *protocol.Request) string {
+	return string(req.Header.Peek(LastEventIDHeader))
+}
+
+// SetLastEventID sets the Last-Event-ID header.
+func SetLastEventID(req *protocol.Request, id string) {
+	req.Header.Set(LastEventIDHeader, id)
+}
+
+// AddAcceptMIME adds `text/event-stream` to http `Accept` header.
+//
+// This is NOT required as per spec:
+// * User agents MAY set (`Accept`, `text/event-stream`) in request's header list.
+func AddAcceptMIME(req *protocol.Request) {
+	v := req.Header.Peek("Accept")
+	if len(v) > 0 {
+		if bytes.Contains(v, bytestr.MIMETextEventStream) {
+			return
+		}
+		// for better compatibility, only use one Accept header value
+		// append `text/event-stream` to the end of the value
+		req.Header.Set("Accept", string(v)+", "+string(bytestr.MIMETextEventStream))
+	} else {
+		req.Header.Set("Accept", string(bytestr.MIMETextEventStream))
+	}
 }
 
 func sseEventType(v []byte) string {
